+++ packages/zarm/src/picker-view/index.ts
@@ -1,6 +1,6 @@
 import PickerView from './PickerView';
 
-export type { PickerDataSource } from './PropsType';
+export type { PickerDataSource } from './interface';
 export type { PickerViewProps } from './PickerView';
 
 export default PickerView;
